fix training

+++ torchreid/integration/sc/train_task.py
@@ -90,7 +90,7 @@ class OTEClassificationTrainingTask(OTEClassificationInferenceTask, ITrainingTas
             update_progress_callback = default_progress_callback
 
         if self._multilabel:
-            self._cfg.train.lr = int(self._cfg.train.lr / self._cfg.sc_integration.lr_scale)
+            self._cfg.train.lr = self._cfg.train.lr / self._cfg.sc_integration.lr_scale
             self._cfg.train.max_epoch = int(self._cfg.train.max_epoch / self._cfg.sc_integration.epoch_scale)
 
         time_monitor = TrainingProgressCallback(update_progress_callback, num_epoch=self._cfg.train.max_epoch,
